Split ContainerTest.php by feature area (ES-239)
